requirements.txt
paramiko==2.0.2
requests==2.10.0
PyYAML==3.12
